test: split oauth test server helper
