+++ e2e/helpers/oauth-test-server/credit-fixtures.ts
@@ -1,0 +1,5 @@
+const creditPurchaseId = `city_paypal_${'ef'.repeat(16)}`
+const creditGiftId = `city_gift_${'ab'.repeat(16)}`
+const creditClaimToken = `gift_claim_${'cd'.repeat(32)}`
+
+export { creditPurchaseId, creditGiftId, creditClaimToken }
